+++ magenta/music/encoder_decoder.py
@@ -37,20 +37,6 @@ OneHotEncoding of individual events. The input vectors are one-hot encodings of
 the most recent event. The output labels are one-hot encodings of the next
 event.
 
-ModuloEventSequenceEncoderDecoder is an EventSequenceEncoderDecoder that uses
-Modulo Encoding for individual note-on events. This encoding can be
-used in place of OneHotEventSequenceEncoderDecoder in models that only use
-note-on and note-off events. The input vectors are modulo-12 encodings
-of the most recent note event. The output labels are one-hot encodings of the
-next note event.
-
-ModuloPerformanceEventSequenceEncoderDecoder is an EventSequenceEncoderDecoder
-that uses Modulo Encoding for individual events. This encoding can be used in
-place of OneHotEventSequenceEncoderDecoder in models that use time-shift and/or
-velocity events in addition to note-on/note-off events. The input vectors are
-modulo-12 encodings of the most recent event. The output labels are one-hot
-encodings of the next event.
-
 LookbackEventSequenceEncoderDecoder is an EventSequenceEncoderDecoder that also
 uses a OneHotEncoding of individual events. However, its input and output
 encodings also consider whether the event sequence is repeating, and the input
@@ -62,7 +48,6 @@ from __future__ import division
 from __future__ import print_function
 
 import abc
-import math
 
 # internal imports
 
@@ -72,7 +57,6 @@ import tensorflow as tf
 
 from magenta.common import sequence_example_lib
 from magenta.music import constants
-from magenta.music import performance_lib
 from magenta.pipelines import pipeline
 
 
@@ -357,250 +341,6 @@ class EventSequenceEncoderDecoder(object):
         loglik += np.log(softmax[i][softmax_pos][index])
       all_loglik.append(loglik)
     return all_loglik
-
-
-class ModuloPerformanceEventSequenceEncoderDecoder(EventSequenceEncoderDecoder):
-  """An EventSequenceEncoderDecoder for modulo encoding performance events.
-
-  The only method of this class that is different from that of
-  OneHotEventSequenceEncoderDecoder is events_to_inputs().
-  """
-
-  def __init__(self, modulo_encoding):
-    """Initialize a ModuloPerformanceEventSequenceEncoderDecoder object.
-
-    Args:
-      modulo_encoding: A ModuloEncoding object that transforms events to and
-          from integer indices using modulo encoding.
-    """
-    self._modulo_encoding = modulo_encoding
-
-  @property
-  def input_size(self):
-    return self._modulo_encoding.input_size
-
-  @property
-  def num_classes(self):
-    return self._modulo_encoding.num_classes
-
-  @property
-  def default_event_label(self):
-    return self._modulo_encoding.encode_event(
-        self._modulo_encoding.default_event)
-
-  def events_to_input(self, events, position):
-    """Returns the input vector for the given position in the event sequence.
-
-    Returns a modulo encoding for the given position in the performance event
-      sequence, by modulo-encoding the output of the one-hot-encoder applied to
-      that event position.
-
-    Args:
-      events: A list-like sequence of events.
-      position: An integer event position in the event sequence.
-
-    Returns:
-      An input vector, a list of floats.
-    """
-    input_ = [0.0] * self.input_size
-    offset, _, event_type, value, bins = (self._modulo_encoding
-                                          .encode_modulo_event(
-                                              events[position]))
-    input_[offset] = 1.0  # valid bit for the event
-    offset += 1
-    angle = 0.0
-    if (event_type == performance_lib.PerformanceEvent.NOTE_ON or
-        event_type == performance_lib.PerformanceEvent.NOTE_OFF):
-
-      # Encode the octave of the note.
-      angle = (float(value) * math.pi) / 72.0  # 12 octaves, 144 notes
-      input_[offset] = math.cos(angle)
-      input_[offset + 1] = math.sin(angle)
-      offset += 2
-
-      # Encode the note itself, using the encoder's lookup table.
-      value %= 12
-      cosine = self._modulo_encoding.embed_note(value)[0]
-      sine = self._modulo_encoding.embed_note(value)[1]
-      input_[offset] = cosine
-      input_[offset + 1] = sine
-    else:
-      # This must be a velocity, or a time-shift event. Compute its
-      # modulo-bins embedding.
-      angle = (float(value) * 2.0 * math.pi) / float(bins)
-      input_[offset] = math.cos(angle)
-      input_[offset + 1] = math.sin(angle)
-    return input_
-
-  def events_to_label(self, events, position):
-    """Returns the label for the given position in the event sequence.
-
-    Returns the zero-based index value for the given position in the event
-    sequence, as determined by the one hot encoding.
-
-    Args:
-      events: A list-like sequence of events.
-      position: An integer event position in the event sequence.
-
-    Returns:
-      A label, an integer.
-    """
-    return self._modulo_encoding.encode_event(events[position])
-
-  def class_index_to_event(self, class_index, events):
-    """Returns the event for the given class index.
-
-    This is the reverse process of the self.events_to_label method.
-
-    Args:
-      class_index: An integer in the range [0, self.num_classes).
-      events: A list-like sequence of events. This object is not used in this
-          implementation.
-
-    Returns:
-      An event value.
-    """
-    return self._modulo_encoding.decode_event(class_index)
-
-  def event_to_num_steps(self, unused_event):
-    """Returns the number of time steps corresponding to an event value.
-
-    This is used for normalization when computing metrics. Subclasses with
-    variable step size should override this method.
-
-    Args:
-      unused_event: An event value for which to return the number of steps.
-
-    Returns:
-      The number of steps corresponding to the given event value, defaulting to
-      one.
-    """
-    return 1
-
-
-class ModuloEventSequenceEncoderDecoder(EventSequenceEncoderDecoder):
-  """An EventSequenceEncoderDecoder producing modulo encodings for note events.
-
-  An EventSequenceEncoderDecoder that produces modulo encodings for note events
-  by first encoding them using a one-hot encoder. This can be used for
-  non-performance models, such as melody_rnn model, to make them use modulo
-  encoding. The only method of this class that is different from that of
-  OneHotEventSequenceEncoderDecoder is events_to_inputs().
-  """
-
-  def __init__(self, one_hot_encoding):
-    """Initialize a ModuloEventSequenceEncoderDecoder object.
-
-    Args:
-      one_hot_encoding: A OneHotEncoding object that transforms events to and
-          from integer indices.
-    """
-    self._one_hot_encoding = one_hot_encoding
-
-  @property
-  def input_size(self):
-    # Use 6 floats to represent the following:
-    # no_event, off_event, octave_cos, octave_sin, note_cos, note_sin
-    return 6
-
-  @property
-  def num_classes(self):
-    return self._one_hot_encoding.num_classes
-
-  @property
-  def default_event_label(self):
-    return self._one_hot_encoding.encode_event(
-        self._one_hot_encoding.default_event)
-
-  def events_to_input(self, events, position):
-    """Returns the input vector for the given position in the event sequence.
-
-    Returns a modulo encoding for the given position in the event sequence, by
-    modulo-encoding the output of the one-hot-encoder applied to that event
-    position.
-
-    Args:
-      events: A list-like sequence of events.
-      position: An integer event position in the event sequence.
-
-    Returns:
-      An input vector, a list of floats.
-    """
-    input_ = [0.0] * self.input_size
-    # First find the one-hot-encoding of the event
-    # (i.e., the position of 1 in the one-hot vector), and then use it to
-    # generate a modulo encoding for the event.
-    label = self._one_hot_encoding.encode_event(events[position])
-    if label == 0:
-      input_[0] = 1.0  # no_event
-    elif label == 1:
-      input_[1] = 1.0  # off_event
-    else:
-      # Note: here the assumption is that NUM_SPECIAL_MELODY_EVENTS is 2, and
-      # that is why in the above, the values of 0 and 1 are assumed to carry
-      # no_event and off_event information, respectively, and any other value
-      # is assumed to carry an on_event information when adjusted for the first
-      # two values. The following code needs to be modified if in the future
-      # these assumptions do not hold.
-      label -= constants.NUM_SPECIAL_MELODY_EVENTS
-
-      # Encode the octave of the note.
-      angle = (float(label) * math.pi) / 72.0  # 12 octaves, 144 notes
-      input_[2] = math.cos(angle)
-      input_[3] = math.sin(angle)
-
-      # Encode the note itself, using modulo-12 encoding.
-      angle = (float(label) * math.pi) / 6.0  # 12 notes per octave
-      input_[4] = math.cos(angle)
-      input_[5] = math.sin(angle)
-    return input_
-
-  def events_to_label(self, events, position):
-    """Returns the label for the given position in the event sequence.
-
-    Returns the zero-based index value for the given position in the event
-    sequence, as determined by the one hot encoding.
-
-    Args:
-      events: A list-like sequence of events.
-      position: An integer event position in the event sequence.
-
-    Returns:
-      A label, an integer.
-    """
-    return self._one_hot_encoding.encode_event(events[position])
-
-  def class_index_to_event(self, class_index, events):
-    """Returns the event for the given class index.
-
-    This is the reverse process of the self.events_to_label method.
-
-    Args:
-      class_index: An integer in the range [0, self.num_classes).
-      events: A list-like sequence of events. This object is not used in this
-          implementation.
-
-    Returns:
-      An event value.
-    """
-    return self._one_hot_encoding.decode_event(class_index)
-
-  def labels_to_num_steps(self, labels):
-    """Returns the total number of time steps for a sequence of class labels.
-
-    Args:
-      labels: A list-like sequence of integers in the range
-          [0, self.num_classes).
-
-    Returns:
-      The total number of time steps for the label sequence, as determined by
-      the one-hot encoding.
-    """
-    events = []
-    for label in labels:
-      events.append(self.class_index_to_event(label, events))
-    return sum(self._one_hot_encoding.event_to_num_steps(event)
-               for event in events)
 
 
 class OneHotEventSequenceEncoderDecoder(EventSequenceEncoderDecoder):
