+++ magenta/music/encoder_decoder_test.py
@@ -17,22 +17,13 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from math import cos
-from math import pi
-from math import sin
-
 # internal imports
 import numpy as np
 import tensorflow as tf
 
 from magenta.common import sequence_example_lib
-from magenta.music import constants
 from magenta.music import encoder_decoder
-from magenta.music import performance_lib
 from magenta.music import testing_lib
-from magenta.music.melody_encoder_decoder import MelodyOneHotEncoding
-from magenta.music.performance_encoder_decoder import PerformanceModuloEncoding
-from magenta.music.performance_lib import PerformanceEvent
 
 
 class OneHotEventSequenceEncoderDecoderTest(tf.test.TestCase):
@@ -401,217 +392,6 @@ class MultipleEventSequenceEncoderTest(tf.test.TestCase):
     self.assertEqual(
         [1.0, 0.0, 1.0, 0.0, 0.0],
         self.enc.events_to_input(events, 4))
-
-
-class ModuloPerformanceEventSequenceEncoderTest(tf.test.TestCase):
-  """Test class for ModuloPerformanceEventSequenceEncoder.
-
-  ModuloPerformanceEventSequenceEncoderDecoder is tightly coupled with the
-  PerformanceModuloEncoding class. As a result, in the test set up, the test
-  object is initialized with a PerformanceModuloEncoding object, and tested
-  accordingly. Since this class mainly modifies the input encoding of events
-  and otherwise acts very similarly to the OneHotEventSequenceEncoderDecoder,
-  the events_to_labels(), and class_index_to_event() methods of the class are
-  not tested.
-  """
-
-  def setUp(self):
-    self._num_velocity_bins = 32
-    self._max_shift_steps = 32
-    self.enc = encoder_decoder.ModuloPerformanceEventSequenceEncoderDecoder(
-        PerformanceModuloEncoding(
-            num_velocity_bins=self._num_velocity_bins,
-            max_shift_steps=self._max_shift_steps))
-
-    self._pitch_encoder_width = 5
-    self._velocity_encoder_width = 3
-    self._time_shift_encoder_width = 3
-    self._expected_input_size = (2 * self._pitch_encoder_width +
-                                 self._velocity_encoder_width +
-                                 self._time_shift_encoder_width)
-
-    self._expected_num_classes = (self._num_velocity_bins +
-                                  self._max_shift_steps +
-                                  2 * (performance_lib.MAX_MIDI_PITCH -
-                                       performance_lib.MIN_MIDI_PITCH + 1))
-
-  def testInputSize(self):
-    self.assertEquals(self._expected_input_size, self.enc.input_size)
-
-  def testNumClasses(self):
-    self.assertEqual(self._expected_num_classes, self.enc.num_classes)
-
-  def testDefaultEventLabel(self):
-    label = self._expected_num_classes - self._num_velocity_bins - 1
-    self.assertEquals(label, self.enc.default_event_label)
-
-  def testEventToNumSteps(self):
-    self.assertEquals(self.enc.event_to_num_steps(1), 1)
-
-  def testEventsToInput(self):
-    num_shift_bins = self._max_shift_steps
-    num_velocity_bins = self._num_velocity_bins
-    slow_base = 2.0 * pi / 144.0
-    fast_base = 2.0 * pi / 12.0
-    shift_base = 2.0 * pi / num_shift_bins
-    velocity_base = 2.0 * pi / num_velocity_bins
-
-    expected_pairs = [
-        (PerformanceEvent(event_type=PerformanceEvent.NOTE_ON, event_value=60),
-         [1.0, cos(60.0 * slow_base), sin(60.0 * slow_base),
-          cos(60.0 * fast_base), sin(60.0 * fast_base),
-          0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]),
-        (PerformanceEvent(event_type=PerformanceEvent.NOTE_ON, event_value=0),
-         [1.0, cos(0.0 * slow_base), sin(0.0 * slow_base),
-          cos(0.0 * fast_base), sin(0.0 * fast_base),
-          0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]),
-        (PerformanceEvent(event_type=PerformanceEvent.NOTE_ON, event_value=127),
-         [1.0, cos(127.0 * slow_base), sin(127.0 * slow_base),
-          cos(127.0 * fast_base), sin(127.0 * fast_base),
-          0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]),
-        (PerformanceEvent(event_type=PerformanceEvent.NOTE_OFF, event_value=72),
-         [0.0, 0.0, 0.0, 0.0, 0.0, 1.0,
-          cos(72.0 * slow_base), sin(72.0 * slow_base),
-          cos(72.0 * fast_base), sin(72.0 * fast_base),
-          0.0, 0.0, 0.0, 0.0, 0.0, 0.0]),
-        (PerformanceEvent(event_type=PerformanceEvent.NOTE_OFF, event_value=0),
-         [0.0, 0.0, 0.0, 0.0, 0.0, 1.0,
-          cos(0.0 * slow_base), sin(0.0 * slow_base),
-          cos(0.0 * fast_base), sin(0.0 * fast_base),
-          0.0, 0.0, 0.0, 0.0, 0.0, 0.0]),
-        (PerformanceEvent(
-            event_type=PerformanceEvent.NOTE_OFF, event_value=127),
-         [0.0, 0.0, 0.0, 0.0, 0.0, 1.0,
-          cos(127.0 * slow_base), sin(127.0 * slow_base),
-          cos(127.0 * fast_base), sin(127.0 * fast_base),
-          0.0, 0.0, 0.0, 0.0, 0.0, 0.0]),
-        (PerformanceEvent(
-            event_type=PerformanceEvent.TIME_SHIFT, event_value=10),
-         [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
-          1.0, cos(9.0 * shift_base), sin(9.0 * shift_base),
-          0.0, 0.0, 0.0]),
-        (PerformanceEvent(
-            event_type=PerformanceEvent.TIME_SHIFT, event_value=1),
-         [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
-          1.0, cos(0.0 * shift_base), sin(0.0 * shift_base),
-          0.0, 0.0, 0.0]),
-        (PerformanceEvent(
-            event_type=PerformanceEvent.TIME_SHIFT, event_value=100),
-         [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
-          1.0, cos(99.0 * shift_base), sin(99.0 * shift_base),
-          0.0, 0.0, 0.0]),
-        (PerformanceEvent(event_type=PerformanceEvent.VELOCITY, event_value=5),
-         [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
-          0.0, 0.0, 0.0,
-          1.0, cos(4.0 * velocity_base), sin(4.0 * velocity_base)]),
-        (PerformanceEvent(event_type=PerformanceEvent.VELOCITY, event_value=1),
-         [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
-          0.0, 0.0, 0.0,
-          1.0, cos(0.0 * velocity_base), sin(0.0 * velocity_base)]),
-        (PerformanceEvent(event_type=PerformanceEvent.VELOCITY, event_value=16),
-         [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
-          0.0, 0.0, 0.0,
-          1.0, cos(15.0 * velocity_base), sin(15.0 * velocity_base)]),
-    ]
-
-    events = []
-    position = 0
-    for event, expected_encoded_modulo_event in expected_pairs:
-      events += [event]
-      actual_encoded_modulo_event = self.enc.events_to_input(events, position)
-      position += 1
-      for i in range(self._expected_input_size):
-        self.assertAlmostEqual(expected_encoded_modulo_event[i],
-                               actual_encoded_modulo_event[i])
-
-
-class ModuloEventSequenceEncoderTest(tf.test.TestCase):
-  """Test class for ModuloEventSequenceEncoder.
-
-  The ModuloEventSequenceEncoderDecoder is expected to be used with the
-  MelodyOneHotEncoding object. As such, in the test set up, the test object is
-  initialized with a MelodyOneHotEncoding object, and tested accordingly. Since
-  this class mainly modifies the input encoding of events
-  and otherwise acts very similarly to the OneHotEventSequenceEncoderDecoder,
-  the events_to_labels(), and class_index_to_event() methods of this class are
-  not tested.
-  """
-
-  def setUp(self):
-    self._min_note = 1
-    self._max_note = 60  # max_note is excluded
-    self.enc = encoder_decoder.ModuloEventSequenceEncoderDecoder(
-        MelodyOneHotEncoding(self._min_note, self._max_note))
-    self._expected_input_size = constants.NUM_SPECIAL_MELODY_EVENTS + 4
-    self._expected_num_classes = self._max_note - self._min_note + 2
-
-  def testInputSize(self):
-    self.assertEquals(self._expected_input_size, self.enc.input_size)
-
-  def testNumClasses(self):
-    self.assertEqual(self._expected_num_classes, self.enc.num_classes)
-
-  def testDefaultEventLabel(self):
-    self.assertAllEqual(0, self.enc.default_event_label)
-
-  def testLabelsToNumSteps(self):
-    labels = [0, 1, 0, 2, 0]
-    self.assertEqual(5, self.enc.labels_to_num_steps(labels))
-
-  def testEventsToInput(self):
-    # Create a list containing all possible events ordered by event number,
-    # starting with no-event=-2, note-off=-1, and first-note=self._min_note.
-    events = (range(self._expected_num_classes) -
-              np.ones(self._expected_num_classes) * 2)
-    events[2:] += self._min_note
-
-    slow_base = 2.0 * pi / 144.0
-    fast_base = 2.0 * pi / 12.0
-
-    # A list of pairs of positions in the above events list, and the expected
-    # encoding for the event in that position. The list include the special
-    # events, the first 12 notes, and a sample of notes in higher octaves.
-    expected_pairs = [
-        (0, [1.0, 0.0, 0.0, 0.0, 0.0, 0.0]),  # no-event
-        (1, [0.0, 1.0, 0.0, 0.0, 0.0, 0.0]),  # note-off
-        (2, [0.0, 0.0, cos(0.0 * slow_base), sin(0.0 * slow_base),  # min-note
-             cos(0.0 * fast_base), sin(0.0 * fast_base)]),
-        (3, [0.0, 0.0, cos(1.0 * slow_base), sin(1.0 * slow_base),
-             cos(1.0 * fast_base), sin(1.0 * fast_base)]),
-        (4, [0.0, 0.0, cos(2.0 * slow_base), sin(2.0 * slow_base),
-             cos(2.0 * fast_base), sin(2.0 * fast_base)]),
-        (5, [0.0, 0.0, cos(3.0 * slow_base), sin(3.0 * slow_base),
-             cos(3.0 * fast_base), sin(3.0 * fast_base)]),
-        (6, [0.0, 0.0, cos(4.0 * slow_base), sin(4.0 * slow_base),
-             cos(4.0 * fast_base), sin(4.0 * fast_base)]),
-        (7, [0.0, 0.0, cos(5.0 * slow_base), sin(5.0 * slow_base),
-             cos(5.0 * fast_base), sin(5.0 * fast_base)]),
-        (8, [0.0, 0.0, cos(6.0 * slow_base), sin(6.0 * slow_base),
-             cos(6.0 * fast_base), sin(6.0 * fast_base)]),
-        (9, [0.0, 0.0, cos(7.0 * slow_base), sin(7.0 * slow_base),
-             cos(7.0 * fast_base), sin(7.0 * fast_base)]),
-        (10, [0.0, 0.0, cos(8.0 * slow_base), sin(8.0 * slow_base),
-              cos(8.0 * fast_base), sin(8.0 * fast_base)]),
-        (11, [0.0, 0.0, cos(9.0 * slow_base), sin(9.0 * slow_base),
-              cos(9.0 * fast_base), sin(9.0 * fast_base)]),
-        (12, [0.0, 0.0, cos(10.0 * slow_base), sin(10.0 * slow_base),
-              cos(10.0 * fast_base), sin(10.0 * fast_base)]),
-        (13, [0.0, 0.0, cos(11.0 * slow_base), sin(11.0 * slow_base),
-              cos(11.0 * fast_base), sin(11.0 * fast_base)]),  # min_note+11
-        (19, [0.0, 0.0, cos(17.0 * slow_base), sin(17.0 * slow_base),
-              cos(17.0 * fast_base), sin(17.0 * fast_base)]),
-        (30, [0.0, 0.0, cos(28.0 * slow_base), sin(28.0 * slow_base),
-              cos(28.0 * fast_base), sin(28.0 * fast_base)]),
-        (41, [0.0, 0.0, cos(39.0 * slow_base), sin(39.0 * slow_base),
-              cos(39.0 * fast_base), sin(39.0 * fast_base)]),
-        (60, [0.0, 0.0, cos(58.0 * slow_base), sin(58.0 * slow_base),
-              cos(58.0 * fast_base), sin(58.0 * fast_base)])]
-
-    for position, expected_encoded_modulo_event in expected_pairs:
-      actual_encoded_modulo_event = self.enc.events_to_input(events, position)
-      for i in range(self._expected_input_size):
-        self.assertAlmostEqual(expected_encoded_modulo_event[i],
-                               actual_encoded_modulo_event[i])
 
 
 if __name__ == '__main__':
